增加函数在字节码文件中的表示，修改Function结构体的字段

+++ src/function.h
@@ -4,7 +4,9 @@
 #include <string>
 #include <vector>
 
+// 格式：[function name length:16][function name:var][instruction
+// count:32][instructions:var]
 struct Function {
   std::string name;
-  std::vector<std::byte> instructions;
+  std::vector<std::uint32_t> instructions;
 };
